requirement.txt
asgiref==3.3.4
Django==3.2.4
django-js-asset==1.2.2
django-mptt==0.12.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
